requirements.txt
numpy
matplotlib
PyYAML
opencv-python
scikit-learn
